+++ spotifybackup/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from . import main
+
+
+sys.exit(main())
